VER3-143 remove redundant routes, refactor views to api, add route for client side logging
